config: Enable doze mode
Modern apps without GMS push can handle doze by asking the user to
whitelist them from battery optimization, so there's no reason to not
enable this.

AOSPA have this enabled under vendor/google/gms, but also enable it
here for those not building GMS and to not miss it on future releases.

Change-Id: I84efe5abe2b8c209202f4fd14217d4dfa7f34de1
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>