Hello Uvite!
Starting in our Android 14 based and after, there won't be more
numbering until stable builds.

Alpha, beta and unofficial builds will be differentiated by the
build date.

Stable builds will now be structured using the next scheme, X.Y.Z

X will be the major version, which will grow up every time we update the platform SPL.

Y will be based on the week the build is released, which will be incremental for every week after X has changed.

Z will be used for hotfixes, which will be incremented depending on how many updates the device get within the same week.

With this new scheme users will be able to follow the changelog for each version as it will be posted in our an specific section on the website, and also will help us showing a platform changelog in the new OTA app on each update.

And as always, it's time for the new wallpaper for Paranoid Android Uvite Alpha.

Change-Id: I0056444100f85d7a404896438719a15aa6b3ffeb
Co-authored-by: Fabian Leutenegger <fabian.leutenegger@bluewin.ch>
Signed-off-by: Hernán Castañón Álvarez <herna@paranoidandroid.co>